move networkversion to client, copy keypart to client
