+++ cmd/client/cmd/root.go
@@ -8,9 +8,9 @@ package cmd
 import (
 	"fmt"
 	"os"
-
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
+	comm "github.com/seeleteam/go-seele/cmd/comm"
 )
 
 var rpcAddr string
@@ -37,6 +37,7 @@ func Execute() {
 func init() {
 	cobra.OnInitialize(initConfig)
 	rootCmd.PersistentFlags().StringVarP(&rpcAddr, "addr", "a", "127.0.0.1:55027", "rpc address")
+	rootCmd.AddCommand(comm.GetGenerateKeyPairCmd("client"))
 }
 
 // initConfig reads in the config file and ENV variables if set.
