+++ cmd/node/cmd/root.go
@@ -10,6 +10,7 @@ import (
 	"os"
 
 	"github.com/spf13/cobra"
+	"github.com/seeleteam/go-seele/cmd/comm"
 )
 
 var rpcAddr string
@@ -35,4 +36,5 @@ func Execute() {
 
 func init() {
 	rootCmd.PersistentFlags().StringVarP(&rpcAddr, "addr", "a", "127.0.0.1:55027", "rpc address")
+	rootCmd.AddCommand(comm.GetGenerateKeyPairCmd("node"))
 }
